vim: Fix redraw glitches
I've been seeing redraw glitches with wrapped lines[0]. I noticed that
switching to an ASCII character for `'showbreak'` fixes it, but also
does switching to other Unicode characters other than the ill-fated
"ARROW PONTING DOWNWARDS THEN CURVING RIGHTWARDS".

I confirmed that the codepoint in the file is correct with `ga` and
`g8`, and also that the `file` tool reports the file as being in UTF-8.

I tried a few other suspicions out, like `'noemoji'` and
`'nolazyredraw'` and `'fasttty'` to no avail. Let's just switch this
out.

[0]: For an example, see this video: https://youtu.be/1vZmJa0Xi1Q