hammerspoon: add a bunch of debug logging
Have noticed one issue where chorded caps lock works but chorded return
does not. Added this logging and noticed that logging stopped when the
bug was manifesting. Chorded caps lock only works because it's control
anyway; the backspace functionality doesn't work.

Next step will be to add more hooks and logging to see whether the tap
is dead and needs to be restarted or what.