vim: Set up shortcut variables for "hash -d" directories
This is a bit of a hack, but the motivation here is to be able to do
things like:

```
:Ack foobar $rrn
```

To scope a search to a directory for which I have defined a shortcut
with `hash -d`.

It's a hack because I am sourcing the specific file where these
shortcuts are defined rather than just doing:

- `zsh -c "hash -d"` (startup files don't get evaluated, so nothing
  is found).
- `zsh -i -c "hash -d"` (slow, takes about .25s, and output includes
  unwanted "stty: stdin isn't a terminal" garbage).

Other alternatives:

- Write out `hash -d` to a file at shell start-up time (meh).
- Grep for `hash -d` lines under `~/.zsh` (meh).
- Do the `zsh -i` thing anyway, and mitigate the slowness by doing it
when idle (via a `CursorHold` autocmd that tears itself down after
running).

Whatevs. The chosen method seems simple, reliable, and fast enough.