vim: adjust 'updatetime' to 2000ms
The default here is 4000ms. Prior to the last commit, we would eagerly
load YCM, which would then set it to 2000ms. But now we are waiting
until Vim has been idle for at least 4000ms before loading YCM. This may
be too long. So, let's make 'updatetime' 2000ms, in agreement with what
YCM would have done anyway, which will mean that it is more likely to be
ready for us when we want it to be.