vim: Defer set-up of $shortcut variables, saving 20ms
This was expensive because of the shell out to the `zsh` process.

Note that I've made the "idleboot" mechanism more decoupled now. Now,
you can just register for the "WincentDefer" User autocommand from
anywhere in order to be called once booted and idle.

The names here could probably use some work, so might follow up with a
change if I can think of something better.

With this, we're down to about 150ms to startup (down from 500ms earlier
this evening), and there is not much low-hanging fruit left.