mutt: add $my_account variable
Again, for easier potential switching in the future. But as I prepared
this, I probably won't want this, because any other account I add won't
be a Gmail one and will wind up with totally different conventions
anyway.

Also, in testing, I not that my recent change to shortcuts results in
some less than readable messages of the form "Switching to !" and so on.
Not sure whether I will keep that after all.