mutt: partial refactor of filtering
Confession: haven't even run this, but I wanted to show it as an
intermediate step. The idea is for each rule, start selecting from
scratch (caching should make this fast), so as to avoid issues with
latter rules operating on messages that have already been archived.

I want to change the "DSL" though, so will do that in the next commit
and actually test the darn thing.