mutt: perf-tune some settings
Some of the time, I am seeing suspicious lags in response to keyboard
input. Example (animated gif):

https://cloud.githubusercontent.com/assets/7074/21078055/cce0d678-bf15-11e6-9451-f861dcb80404.gif

In this, you can see me using `j` and `k` to move up and down an index
listing. Occasionally, there is a lag between the input (visible thanks
to KeyCastr) and the action; this shows up as a white cursor next to the
arrow (which I turned on with `:set arrow_cursor` to demo this).

One guess: could it be something to do with mail checking? `set
check_new = no` would be pretty heavy-handed, so instead, let's just up
some intervals:

- `mail_check` goes up to 60s (from the default of 5s).
- `mail_check_stats_interval` goes up to 300s (from the default of 60s).

And we also set maildir_header_cache_verify to no, which the man page
suggests should be ok (it talks about maildir-unaware programs modifying
the messages, but I don't think offlineimap qualifies as a
"maildir-unaware" program).