mutt: update lbdb on postsync
Took the opportunity to refactor the postsync a little as well, seeing
as we have to run two things. Use `lockrun` to ensure that we don't run
`notmuch` more than once at a time. (Could also have used `--wait`
instead of `--quiet`, but it will run again on the next sync anyway,
which will be about a minute later.)