mutt: add control.sh process for pausing syncing
Convenience for pausing syncing when on a shitty network. There are
other ways that this could have been done, such as teaching "sync.sh"
itself to look for marker files after each operation, but this gives us
the ability to do an immediate SIGSTOP even in the middle of an
operation, and will also give me a place to put the launch invocation
(in case it is not in the command history).