mutt: tweak in an attempt to avoid long hangs
Suspect here is `$mail_check_stats`, given the amount of mail I have.

Suggestion to also "increase" `$timeout` from here:

http://marc.info/?l=mutt-users&m=148604958727065&w=2

("Increasing" it from 0 to 60 here actually means changing it from
infinity to 60.)

The only drawback here is that without `$mail_check_stats` I can no
longer reliably show the message count in the folder list, so I am
removing that. Additionally the count in the sidebar is missing (0) or
inaccurate until I go to that mailbox, but if this solves the hangs then
I think it will be worth it. Giving it a try anyway. If the inaccuracy
becomes too annoying, I'll see if I can live without the count in the
sidebar.