mutt: suppress appearance of unwanted Archive1 folder
Looks like an Outlook bug:

https://answers.microsoft.com/en-us/outlook_com/forum/oemail-ofolders/archive1-folder-automatically-appears-in/a2c8d4c3-822e-4166-be8c-1053f854e355

Nothing to do with mbsync (even when mbsync is not running, I can delete
the folder in the Outlook web UI, and it's back there again on the next
refresh). Currently trying to delete it using the macOS Outlook app, but
in any case, going to implement this band-aid for now so that I won't
have to see the folder no matter what happens.