vim: Update base16-vim
Trying to fix busted syntax highlighting with some template literals.

Note that this requires a tweak because one of the commits here drops
the fg color from the `MatchParen` highlight. We were using that in our
`User5` and `User6` rules, which were causing Vim to freak out with:

    E254: Cannot allocate color guibg=#b4b7b4

because it would try to evaluate these rules:

    highlight User5 gui=bold guifg= guibg=#b4b7b4
    highlight User6 gui=bold,italic guifg= guibg=#b4b7b4

And treat "guibg=#b4b7b4" as the value for `guifg`, which is obviously
gargbage. Pull the color (#1d1f21) off of the `Cursor` highlight
instead. Haven't tested this in other schemes than tomorrow-night yet,
as I am barely using anything else nowadays, but I will check a few
faves.

* roles/dotfiles/files/.vim/pack/bundle/start/base16-vim 9bc7051...02955c9 (12):
  > Merge pull request #140 from anuragsoni/update-color-schemes
  > Merge pull request #137 from tmccombs/matchparen
  > Merge pull request #136 from houjunchen/vim-plug
  > Merge pull request #135 from danielrs/master
  > Merge pull request #134 from idhx/master
  > Merge pull request #133 from getkey/pangloss-js
  > Merge pull request #132 from Naissur/master
  > Updates output
  > Brightens color for CursorLineNr
  > Updates output
  > Merge branch 'master' of ssh://github.com/chriskempson/base16-vim
  > Scheme update