mutt: filter GitHub emails sent to work account
Trying to achieve a better separation of work and personal emails, I've
set up my GitHub notifications to go to my personal account for personal
projects, and to my work account for work projects:

https://github.com/blog/1204-notifications-stars

So, here we just copy-pasta the GitHub set-up from `home.lua` to
`work.lua`. The only difference is that `contain_field` is broken on
Exchange, so I had to change that to `match_field`, which is slower, but
at least works.

Now that we have this copy-pasta, I should clean it up. The
`differential_related` code is very like the `github_related` code, so I
should unify those too.