vim: Toggle LHS statusline color when &modified
Decided to do this because my recent change to move the "+" indicator to
the LHS is pretty subtle and hard to see with the colors I'm using.

This is an ugly manually coded state machine (juggling
modified/nomodified vs async active/inactive) but it seems to work.

Some quirks to note:

- In my testing `exists('#TextChanged')` is 0 when Vim first boots but
  `exists('#TextChangedI')` is 1, so I'm using that. No idea why.
- Not clear whether the autocmd set-up in `statusline.vim` would be
  better in `autocmds.vim`, or contrarily, the statusline-related
  autocmds in `autocmds.vim` would be better off in `statusline.vim`.