Link to dependent projects uniformly
All the other projects listed in the dependencies
section have links to the main project pages, not
the download pages. While that is useful, not all
people want to install things via manual download
(many will use a package manager or a tool like
Hombrew). So, let's stay unopinionated and consistent
with what the other entries are doing.