vim: [experimental] turn off syntax highlighting in inactive splits
Closes: https://github.com/wincent/wincent/issues/37

May be too slow on very big screens, even with the viewport scoping of
the `matchaddpos` calls. Will see if this is too flaky or not. In any
case, may want to guard it with some feature detection (not sure when
`matchaddpos` was added, nor support for "w0" and "w$" in `line()`).