tmux: make sure SHLVL always starts at 1
Now that I have SHLVL reflected in the prompt, I want this to be
consistent.

You would expect to have SHLVL 2 in the nested shell inside the tmux
session, however I was seeing SHLVL 3 when invoked via a wrapper script
(like the `.tmux` file in this repo). This was confusing because this
is a global setting, so the level gets determined for all new sessions
based on whether the first session happened to be invoked by a wrapper
or not.

With this change, we always start at 2 in the nested shell, no matter
what.