vim: drop priority on window blur match highlighting
Default priority is 10, `'hlsearch'` priority is 0, and `'syntax'`
highlighting is completely independent (`matchadd()`/`matchaddpos()`
always override it), so, by dropping the pri here to 0 we can still see
search highlights when we switch to other windows.

In using the new focus/blur for almost a day now, I've found this to be
useful. I often search for things and end up jumping around splits, and
wanting to refer back visually to the thing that I searched for in other
panes.

See `:h matchadd()` for more detail on priorities.