vim: copy over match.c from Command-T
Want to use this as a basis for a Python C extension to improve the
ranking of fuzzy matches in deoplete.nvim results. This is a copy from
the current "next" branch, corresponding to commit 5c8645893c9ddc09.

Won't have to do much, I *think*, as this mostly deals with string
pointers and lengths. The hard part will actually be figuring out how
to build, use and distribute C extensions in Python; it's a totally
foreign ecosystem to me. Starting points:

- https://docs.python.org/3/extending/extending.html
- https://docs.python.org/3/distributing/index.html
- https://docs.python.org/3/extending/index.html
- https://docs.python.org/3/c-api/index.html
- https://packaging.python.org/extensions/
- https://pypi.python.org/pypi

Once I've prototyped this out and confirmed it works, will of course
move this into a more suitable location (possibly another repo of its
own, or a branch on the Command-T repo).