zsh: make `color -` return to previous color scheme
Based on and closes: https://github.com/wincent/wincent/pull/31

That one doesn't apply cleanly any more, and the additional fix
(https://github.com/lencioni/dotfiles/commit/249de701) likewise won't
apply, so I just rewrote it from scratch.

Not bothering with the "don't write if same" check, because:

- I don't think I've ever run "color same-scheme-again"; and
- I want to match the behavior of `cd` (where if you do
  `cd a; cd b; cd .; cd -` you'll remain in "b").

Unlike the upstream patch, I phrased this in terms of an anonymous
inner function `__color()` rather than a recursive call to `color()`
because paranoia (wouldn't want anybody to insert a "-" in the previous
config file).

Thanks to @lencioni for the PR on which this is based.