Karabiner: add hacky Left+Right-Shift → Caps Lock mappings
This used to work in the old Karabiner as "tap left and right shift at
the same time get caps lock". With this hack we get something almost as
useful: "hold down left shift then tap right shift to toggle Caps Lock".

This is workable until we get the real deal.