hammerspoon: remove keylogger
Because it was causing problems with Karabiner-Elements:

    https://github.com/tekezo/Karabiner-Elements/issues/789

In any case, I have gathered more than enough "interesting" data from
the logger at this point.

Since adding the module in 02bedf49778f965b3e0f41 (and ancestor commits)
and having it running mostly full-time since April 26 (that is, 57
days), I've collected:

Unique n-grams:

    sqlite> select count(1) from unigrams;
    48
    sqlite> select count(1) from digrams;
    2225
    sqlite> select count(1) from trigrams;
    31280

Total counts of n-grams:

    sqlite> select sum(count) from unigrams;
    1892752
    sqlite> select sum(count) from digrams;
    1610780
    sqlite> select sum(count) from trigrams;
    1419347

So we're looking at a lower bound of 33k keystrokes per day (lower
because we're not counting modifiers or other non-core key-presses),
which, if spread evenly throughout the day would be only one every 2.6
seconds. This feels about "ball-park" right.

The top 10 unigrams are:

    sqlite> select * from unigrams order by count desc limit 10;
     |239755
    j|140941
    e|119745
    k|111869
    t|105163
    n|87786
    i|83302
    a|82225
    o|80160
    s|69764

Happily, neither "j" nor "k" is number one (that would be a bad sign).
Space is the best, and common English language "e" makes a strong
showing in third place, beaten by Vim's "j" for second. The others in
the top 10 are all Colemak home-row keys, as you would expect, with
Vim's "k" grabbing fourth place to be the only other non-home entry.

Here are the top 10 digrams:

    sqlite> select * from digrams order by count desc limit 10;
    jj|94514
    kk|69603
      |27456
    t |22040
    e |21902
     t|19548
     w|18529
    in|17407
    re|16193
    th|15233

Here "jj" and "kk" dominate (the shame!). Double-space makes a strong
showing in third place, which either means I switch buffers a lot (with
"<Leader><Leader>" or I do a lot of paging, or both. The other entries
all look like standard common English digrams with strong Colemak
home-row involvement, except for " w" which is evidence of me
compulsively saving my Vim buffers with "<Leader>w".

Last up, the top 10 trigrams:

    sqlite> select * from trigrams order by count desc limit 10;
    jjj|70198
    kkk|50569
    nnn|10687
     th|9350
       |8975
    it |7259
    bbb|7117
    www|6780
    jjk|6527
    the|6506

Almost all of this is evidence of bad Vim habits, with the exception of
" th" and "the" which are in there because of English. "jjj" and "kkk"
dominate because of lazy vertical movement, "nnn" because that's how you
move between search results, and "bbb" and "www" because of by-word
movement, doubtless some of which should be replaced with "f", "F", "t",
"T" and other alternatives.