git-hub: avoid use of "git-am" in "get" subcommand
When I originally wrote this "git-am" was fine because I was generally
just resurrecting old commits, which needed to be rebased (mutated)
anyway.

In many cases, however, I don't want to mutate the history
unnecessarily. So, switch this to "git-fetch" instead and ditch the
"curl" call. You can always rebase from there if you want to.

Should probably apply the same treatment to the "update" subcommand, but
I don't need that yet, so not doing it.