Karabiner: alt take on SpaceFN layer
This one doesn't really work either (fails to deal with rollover), but
committing it to show use of variables for layers.

Also moved the descriptions so they show up in the UI properly, and
added a "comment" (ie. an ignored "description" property) to remind
myself on what the "Realforce" profile does.