karabiner: make use of device_if to replace separate profiles
Previously I had an "Internal" and a "Realforce" profile which were the
same except the latter swapped Option and Command (to match the internal
Mac keyboard). This involved a bunch of duplication and some custom
Hammerspoon stuff to automate the switching between profiles.

That all worked pretty well, but with `device_if` just getting added to
Karabiner-Elements, I can get rid of the duplication.

Also sneaking in in this commit: disabling the SpaceFN layer because it
doesn't (yet) work too well. We need an equivalent of the old
`__BlockUntilKeyUp__` functionality.

Renamed "Internal" to "Default".

Reworded some descriptions for consistency and to add some missing
detail.