mutt: add "term" command to control script
This is mostly to unstick stuck mbsync processes, which tend to happen
when I lose network connectivity, especially on the Gmail sync. As this
is a nasty operation, it takes an optional parameter to target a
specific process (home or work). For consistency, I made the other
relevant comands take target params.