Ansible: switch to dynamic inventory
I'd like to conditionalize the behavior based on the hostname of the
local host, but Ansible will always use "localhost" so I can't do the
simple thing with a static inventory file that looks like this:

  [local]
  localhost ...

  [work]
  glh-mbp ...

Instead, use the dynamic inventory feature by turning the inventory into
a directory and putting an executable Python file in there which returns
the inventory in as a JSON blob.

All it does is look at the hostname and return JSON containing a
group name of "personal", "work", "devservers" or (as a fallback)
"local". With that in place, we're free to put our variables inside
"group_vars/{personal,work}.yml" files.